requirements.txt
load_dotenv 
openai 
pandas
openpyxl